denie me

+++ src/s21_memchr.c
@@ -1,8 +1,8 @@
 #include "s21_string.h"
 
 void *s21_memchr(const void *str, int pattern, size_t count) {
-  size_t result = -1;
-  for (int i = 0; i < count; i++) {
+  int result = -1;
+  for (size_t i = 0; i < count; i++) {
     if (*(unsigned char *)(str + i) == (unsigned char)pattern) {
       result = i;
       break;
