+++ src/s21_memcmp.c
@@ -1,8 +1,8 @@
 #include "s21_string.h"
 
 int s21_memcmp(const void *s1, const void *s2, size_t n) {
-  int result = 0;
-  for (int i = 0; i < n; i++) {
+  size_t result = 0;
+  for (size_t i = 0; i < n; i++) {
     result = *(unsigned char *)(s1 + i) - *(unsigned char *)(s2 + i);
     if (result != 0) {
       break;
