+++ src/s21_memset.c
@@ -1,7 +1,8 @@
 #include "s21_string.h"
 
+
 void *s21_memset(void *b, int c, size_t len) {
-  for (int i = 0; i < len; i++) {
+  for (size_t i = 0; i < len; i++) {
     *(unsigned char *)(b + i) = (unsigned char)c;
   }
   return (unsigned char *)b;
