+++ src/s21_strncat.c
@@ -4,7 +4,7 @@
 // на которую указывает dest, длиной до n символов.
 
 char* s21_strncat(char* dest, const char* src, size_t n) {
-    int i, j;
+    size_t i, j;
 
     for (i = 0; dest[i] != '\0'; i++);
